GCI-951: Separate out request mapping for reuse.
