+++ src/main/java/uk/gov/companieshouse/orders/api/interceptor/UserAuthenticationInterceptor.java
@@ -1,12 +1,7 @@
 package uk.gov.companieshouse.orders.api.interceptor;
 
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
-import org.springframework.web.servlet.mvc.condition.PatternsRequestCondition;
-import org.springframework.web.servlet.mvc.condition.RequestMethodsRequestCondition;
 import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
@@ -20,61 +15,26 @@ import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 import static uk.gov.companieshouse.orders.api.OrdersApiApplication.APPLICATION_NAMESPACE;
-import static uk.gov.companieshouse.orders.api.controller.BasketController.*;
-import static uk.gov.companieshouse.orders.api.controller.OrderController.GET_ORDER_URI;
+import static uk.gov.companieshouse.orders.api.interceptor.RequestMapper.*;
 import static uk.gov.companieshouse.orders.api.util.EricHeaderHelper.API_KEY_IDENTITY_TYPE;
 import static uk.gov.companieshouse.orders.api.util.EricHeaderHelper.OAUTH2_IDENTITY_TYPE;
 
 @Service
-public class UserAuthenticationInterceptor extends HandlerInterceptorAdapter implements InitializingBean {
+public class UserAuthenticationInterceptor extends HandlerInterceptorAdapter {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(APPLICATION_NAMESPACE);
 
-    static final String ADD_ITEM = "addItem";
-    static final String CHECKOUT_BASKET = "checkoutBasket";
-    static final String GET_PAYMENT_DETAILS = "getPaymentDetails";
-    static final String PATCH_BASKET = "patchBasket";
-    static final String PATCH_PAYMENT_DETAILS = "patchPaymentDetails";
-    static final String GET_ORDER = "getOrder";
+    private final RequestMapper requestMapper;
 
-    private final String addItemUri;
-    private final String checkoutBasketUri;
-    private final String patchBasketUri;
-    private final String getPaymentDetailsUri;
-    private final String getOrderUri;
-    private final String patchPaymentDetailsUri;
-
-    /**
-     * Represents the requests authenticated by this.
-     */
-    private List<RequestMappingInfo> authenticatedRequests;
-
-    public UserAuthenticationInterceptor(
-            @Value(ADD_ITEM_URI)
-            final String addItemUri,
-            @Value(CHECKOUT_BASKET_URI)
-            final String checkoutBasketUri,
-            @Value(PATCH_BASKET_URI)
-            final String patchBasketUri,
-            @Value(GET_PAYMENT_DETAILS_URI)
-            final String getPaymentDetailsUri,
-            @Value(GET_ORDER_URI)
-            final String getOrderUri,
-            @Value(PATCH_PAYMENT_DETAILS_URI)
-            final String patchPaymentDetailsUri) {
-        this.addItemUri = addItemUri;
-        this.checkoutBasketUri = checkoutBasketUri;
-        this.patchBasketUri = patchBasketUri;
-        this.getPaymentDetailsUri = getPaymentDetailsUri;
-        this.getOrderUri = getOrderUri;
-        this.patchPaymentDetailsUri = patchPaymentDetailsUri;
+    public UserAuthenticationInterceptor(final RequestMapper requestMapper) {
+        this.requestMapper = requestMapper;
     }
 
     @Override
     public boolean preHandle(final HttpServletRequest request,
                              final HttpServletResponse response,
                              final Object handler) {
-        final RequestMappingInfo match = getRequestMapping(request);
+        final RequestMappingInfo match = requestMapper.getRequestMapping(request);
         if (match != null) {
             switch (match.getName()) {
                 case ADD_ITEM:
@@ -92,21 +52,6 @@ public class UserAuthenticationInterceptor extends HandlerInterceptorAdapter imp
             }
         }
         return true;
-    }
-
-    /**
-     * Gets the request mapping found for the request provided.
-     * @param request the HTTP request to be authenticated
-     * @return the mapping representing the request if it is to be authenticated, or <code>null</code> if not
-     */
-    RequestMappingInfo getRequestMapping(final HttpServletRequest request) {
-        for (final RequestMappingInfo mapping: authenticatedRequests) {
-            final RequestMappingInfo match = mapping.getMatchingCondition(request);
-            if (match != null) {
-                return match;
-            }
-        }
-        return null; // no match found
     }
 
     /**
@@ -212,42 +157,4 @@ public class UserAuthenticationInterceptor extends HandlerInterceptorAdapter imp
         return true;
     }
 
-    @Override
-    public void afterPropertiesSet() {
-        final RequestMappingInfo addItem =
-                new RequestMappingInfo(ADD_ITEM,
-                        new PatternsRequestCondition(addItemUri),
-                        new RequestMethodsRequestCondition(RequestMethod.POST),
-                        null, null, null, null, null);
-        final RequestMappingInfo checkoutBasket =
-                new RequestMappingInfo(CHECKOUT_BASKET,
-                        new PatternsRequestCondition(checkoutBasketUri),
-                        new RequestMethodsRequestCondition(RequestMethod.POST),
-                        null, null, null, null, null);
-        final RequestMappingInfo getPaymentDetails =
-                new RequestMappingInfo(GET_PAYMENT_DETAILS,
-                        new PatternsRequestCondition(getPaymentDetailsUri),
-                        new RequestMethodsRequestCondition(RequestMethod.GET),
-                        null, null, null, null, null);
-        final RequestMappingInfo patchBasket =
-                new RequestMappingInfo(PATCH_BASKET,
-                        new PatternsRequestCondition(patchBasketUri),
-                        new RequestMethodsRequestCondition(RequestMethod.PATCH),
-                        null, null, null, null, null);
-        final RequestMappingInfo patchPaymentDetails =
-                new RequestMappingInfo(PATCH_PAYMENT_DETAILS,
-                        new PatternsRequestCondition(patchPaymentDetailsUri),
-                        new RequestMethodsRequestCondition(RequestMethod.PATCH),
-                        null, null, null, null, null);
-        final RequestMappingInfo getOrder =
-                new RequestMappingInfo(GET_ORDER,
-                        new PatternsRequestCondition(getOrderUri),
-                        new RequestMethodsRequestCondition(RequestMethod.GET),
-                        null, null, null, null, null);
-
-        authenticatedRequests = asList(
-                addItem, checkoutBasket, getPaymentDetails, patchBasket, patchPaymentDetails, getOrder
-        );
-
-    }
 }
