+++ src/test/java/uk/gov/companieshouse/orders/api/interceptor/UserAuthenticationInterceptorTests.java
@@ -14,10 +14,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
-import static org.hamcrest.core.IsNull.nullValue;
 import static org.mockito.Mockito.*;
 import static org.springframework.http.HttpMethod.*;
-import static uk.gov.companieshouse.orders.api.interceptor.UserAuthenticationInterceptor.*;
 import static uk.gov.companieshouse.orders.api.util.EricHeaderHelper.*;
 import static uk.gov.companieshouse.orders.api.util.TestConstants.ERIC_IDENTITY_INVALID_TYPE_VALUE;
 import static uk.gov.companieshouse.orders.api.util.TestConstants.ERIC_IDENTITY_VALUE;
@@ -272,83 +270,6 @@ public class UserAuthenticationInterceptorTests {
 
         // When and then
         thenRequestIsRejected();
-    }
-
-    @Test
-    @DisplayName("getRequestMappingInfo gets the add item request mapping")
-    void getRequestMappingInfoGetsAddItem() {
-
-        // Given
-        givenRequest(POST, "/basket/items");
-
-        // When and then
-        assertThat(interceptorUnderTest.getRequestMapping(request).getName(), is(ADD_ITEM));
-    }
-
-    @Test
-    @DisplayName("getRequestMappingInfo gets the checkout basket request mapping")
-    void getRequestMappingInfoGetsCheckoutBasket() {
-
-        // Given
-        givenRequest(POST, "/basket/checkouts");
-
-        // When and then
-        assertThat(interceptorUnderTest.getRequestMapping(request).getName(), is(CHECKOUT_BASKET));
-    }
-
-    @Test
-    @DisplayName("getRequestMappingInfo gets the get payment details request mapping")
-    void getRequestMappingInfoGetsGetPaymentDetails() {
-
-        // Given
-        givenRequest(GET, "/basket/checkouts/{checkoutId}/payment");
-
-        // When and then
-        assertThat(interceptorUnderTest.getRequestMapping(request).getName(), is(GET_PAYMENT_DETAILS));
-    }
-
-    @Test
-    @DisplayName("getRequestMappingInfo gets the patch basket request mapping")
-    void getRequestMappingInfoGetsPatchBasket() {
-
-        // Given
-        givenRequest(PATCH, "/basket");
-
-        // When and then
-        assertThat(interceptorUnderTest.getRequestMapping(request).getName(), is(PATCH_BASKET));
-    }
-
-    @Test
-    @DisplayName("getRequestMappingInfo gets the patch payment details request mapping")
-    void getRequestMappingInfoGetsPatchPaymentDetails() {
-
-        // Given
-        givenRequest(PATCH, "/basket/checkouts/{id}/payment");
-
-        // When and then
-        assertThat(interceptorUnderTest.getRequestMapping(request).getName(), is(PATCH_PAYMENT_DETAILS));
-    }
-
-    @Test
-    @DisplayName("getRequestMappingInfo gets the get order request mapping")
-    void getRequestMappingInfoGetsGetOrder() {
-
-        // Given
-        givenRequest(GET, "/orders/{id}");
-
-        // When and then
-        assertThat(interceptorUnderTest.getRequestMapping(request).getName(), is(GET_ORDER));
-    }
-
-    @Test
-    @DisplayName("getRequestMappingInfo returns null where no mapping found")
-    void getRequestMappingInfoReturnsNullWhereNoMappingFound() {
-
-        // Given
-        givenRequest(DELETE, "/unknown/uri");
-
-        // When and then
-        assertThat(interceptorUnderTest.getRequestMapping(request), is(nullValue()));
     }
 
     /**
